[Spring 지하철 노선도 - 3단계] 조조그린(조한석) 미션 제출합니다. (#322)
* feat(StationService): searchById 메서드 구현

* feat: Section 클래스 구현

* feat: 노선이 등록될 때 구간도 저장되는 기능 구현

* refactor: SectionService 를 LineService 로 통합

* feat: Section 이 추가될수 있는지 판별하는 기능 구현

* feat: Section 을 추가하는기능 구현

* feat: Line 을 조회하면 순서대로 정렬된 역을 반환하는 기능 구현

* chore: 줄바꿈

* feat: Section 을 삭제하는 기능 구현

* refactor: Station 을 삭제하는 기능 수정

* test: StationServiceTest 수정

* test: LineServiceTest 수정, 추가

* test: SectionsOnTheLineTest 구현

* test: JdbcLineDaoTest, JdbcStationDaoTest 수정

* test: JdbcSectionDaoTest 구현

* teat: AcceptanceTest 수정, 구현

* docs: README.md 업데이트

* test: JdbcSectionDaoTest 에러 수정

* docs: jdbcSectionDaoSetting.sql 수정

* refactor: id 가 null 인 도메인 객체 생성을 정적팩토리 메서드가 해주도록 변경

* refactor: LineService 의 Line 관련 메서드 리팩토링

* refactor: LineService 의 Section 관련 메서드 리팩토링

* test: 테스트 수정

* refactor(JdbcLineDao): findById 수정

* refactor: SectionsOnTheLine 메서드 순서 변경

* refactor: optimize imports

* test: 불필요한 라인 제거

+++ src/main/java/wooteco/subway/service/StationService.java
@@ -1,12 +1,16 @@
 package wooteco.subway.service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 import org.springframework.dao.DuplicateKeyException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.request.StationRequestDto;
+import wooteco.subway.exception.CanNotDeleteException;
 import wooteco.subway.exception.DuplicateStationNameException;
+import wooteco.subway.repository.dao.SectionDao;
 import wooteco.subway.repository.dao.StationDao;
 import wooteco.subway.repository.entity.StationEntity;
 
@@ -14,29 +18,41 @@ import wooteco.subway.repository.entity.StationEntity;
 public class StationService {
 
     private final StationDao stationDao;
+    private final SectionDao sectionDao;
 
-    public StationService(final StationDao stationDao) {
+    public StationService(final StationDao stationDao, final SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
     public Station register(final StationRequestDto stationRequestDto) {
-        final Station station = new Station(stationRequestDto.getName());
+        final Station station = Station.ofNullId(stationRequestDto.getName());
         try {
-            final StationEntity savedStationEntity = stationDao.save(new StationEntity(station));
-            return savedStationEntity.generateStation();
+            return stationDao.save(new StationEntity(station)).createStation();
         } catch (DuplicateKeyException exception) {
             throw new DuplicateStationNameException();
+        }
+    }
+
+    public Station searchById(final Long id) {
+        try {
+            return stationDao.findById(id).createStation();
+        } catch (EmptyResultDataAccessException exception) {
+            throw new NoSuchElementException("[ERROR] 역을 찾을 수 없습니다.");
         }
     }
 
     public List<Station> searchAll() {
         return stationDao.findAll()
                 .stream()
-                .map(stationEntity -> new Station(stationEntity.getId(), stationEntity.getName()))
+                .map(StationEntity::createStation)
                 .collect(Collectors.toList());
     }
 
     public void remove(final Long id) {
+        if (sectionDao.findByStationId(id).size() != 0) {
+            throw new CanNotDeleteException();
+        }
         stationDao.deleteById(id);
     }
 }
